Fix broken pronouns in vore messages
Okay, BYOND's pronoun text macros are kinda dumb.  Turns out the pronoun
macros "\his" etc operate based on the gender of the PRECEEDING mob
variable.  You can't actually explicitly tell it which variable you want
referenced, and can't suppress the actual output of the variable.
(unlike \ref, which suppresses the normal output)
Therefore we use this sneakyhack!  the unusual  "blah ["[prey]"] blah"
syntax evaluates  "[prey]" into a string, which is then inserted into
the enclosing string.  Because the variable is of type string, it is not
a mob, and therefore looked over when searching for the preceding mob
from \his.