Asteroid Mineral Distribution Overhaul
Minerals on the asteroid consist of the underground resources and the
actual turf itself.  The distribution of surface turf minerals resulted
in very low amounts.
Part of this is because it distributes it randomly throughout the
designated asteroid area, without actually checking if the tile it plans
to put minerals on is actually designated in the map as randomizable
(masked).   When it comes time to actually place the minerals,
non-masked tiles are skipped, so you end up with less minerals than
planned.  This can be solved by simply checking.
The second issue is that the map generator decides to place a bunch of
mineral turfs, but the turf itself then randomly picks what type.   This
does not happen until its instantiated; too late to efficiently try and
reroll.
Instead we have pre-defined numbers of tiles that will be seeded with
minerals.