Shrink & Grow Mushrooms back!
Fixing issue #73, made some sprites and added the logic to the
mushrooms, adding them to the botanics and xenobotany seeds lists. I
also fixed an error of logic in the micro and macro chems, that got
broken during the initial porting.

I also added ready to spawn mushrooms for admins and mappers. megam and
microm.