'Prey' AI Module
Now, what nobody asked for, the 'Prey' AI module.

Based on 'OneHuman', this board specifies one person is food and not a
crew member, and that they must not go to waste.

Spawns with the other balloon crate. So if you get a balloon, you get an
AI board. Consistency!