Rig Correction and Addition
I mistakenly through the 'slowdown' variable for armor was AFTER it had
lost power, not before. So I made the ccmed_rig have high slowdown. It's
almost unusable, even when powered. So, adjusted that.

Secondly, there were unused sprites for security_rig, which I've modified
to fit muzzles and given to the HoS on a rack on their office. It has a
mounted egun. It's not in their locker (following the tradition of rig
modules being on a rack in the head office), but they are unhackable
without an emag.