Updated alt_titles Chaplain
I looked at one of the issues in this github and decided to set on something simple to help out people, the ALT TITLES are the only variable I changed and I added quite a few, understand this is only the file for the chaplain. I will throw the changes down bellow since I haven't used github in a while and I'm unsure if it records any of that.

alt_titles = list("Counselor","Therapist","Preist","Cleric","Divine","Monk","Preistess","Pope")
As above, I added a few for example and might decide to add more if we really need it. I want to look at the other jobs and do the same if this commit gets accepted. - I've also tested this and compiled it.