Adds a new xenoarcheology tool
Adds a tool that incorporates the the entire excavation pick kit and the
hand pick into one, allowing xenoarcheologists to only require one item
to excavate artifacts.

Requires two sheets of iron, 1/2 a sheet of glass, 1/2 a sheet of
silver, and 1/4 a sheet of diamond in order to produce one unit.

Requires the following tech levels to be produced:
"materials" = 6, "powerstorage" = 3, "engineering" = 3, "bluespace" = 4